TST: Split tests/indexes/interval/test_interval.py into separate files
